+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/ConstantIndicator.java
@@ -31,7 +31,6 @@ import org.ta4j.core.indicators.AbstractIndicator;
  */
 public class ConstantIndicator<T> extends AbstractIndicator<T> {
 
-    private static final long serialVersionUID = -186917236870375024L;
     private final T value;
 
     public ConstantIndicator(BarSeries series, T t) {