+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - **Breaking:** **`Order.class`** renamed to **`Trade.class`**
 - **Breaking:** package "tradereports" renamed to "reports"
 - **Breaking:** package "trading/rules" renamed to "rules"
+- **Breaking:** remove Serializable from all indicators
 
 
 ### Fixed
@@ -70,6 +71,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - :tada: **Enhancement** added Chainrule.
 - :tada: **Enhancement** added BarSeriesUtils#sortBars.
 - :tada: **Enhancement** added BarSeriesUtils#addBars.
+- :tada: **Enhancement** added Num.negate() to negate a Num value.
 - :tada: **Enhancement** added **`GrossLossCriterion.class`**.
 - :tada: **Enhancement** added **`NetProfitCriterion.class`**.
 - :tada: **Enhancement** added chooseBest() method with parameter tradeType in AnalysisCriterion.